requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
mysqlclient
Pillow
markdown
django-filter
django-oauth-toolkit
drf-yasg